# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Temple API",
    "description": "Docs for /temples",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/professional": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/temples/": {
      "get": {
        "tags": [
          "Temples"
        ],
        "description": "Return all temples",
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Temples"
        ],
        "description": "Create a new temple",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "temple_id": {
                  "type": "number",
                  "example": 9
                },
                "name": {
                  "type": "string",
                  "example": "New Temple"
                },
                "location": {
                  "type": "string",
                  "example": "Somewhere"
                },
                "dedicated": {
                  "type": "string",
                  "example": "TBD"
                },
                "additionalInfo": {
                  "type": "boolean",
                  "example": false
                }
              },
              "required": [
                "temple_id",
                "name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/temples/{temple_id}": {
      "get": {
        "tags": [
          "Temples"
        ],
        "description": "Return a single temple by temple_id",
        "parameters": [
          {
            "name": "temple_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}